requirements.txt
torch -i https://download.pytorch.org/whl/cu118
torchvision -i https://download.pytorch.org/whl/cu118
numpy==1.24.1
matplotlib==3.7.1
opencv-python==4.7.0.72
onnx==1.13.1
onnxruntime==1.14.1
flask==2.2.3
flask-socketio==5.3.3
socketio-client==0.7.2
pyyaml==6.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git